Flag for setting status code on integrityerrors
If an SQLAlchemy IntegrityError is caught during a commit a 400 will be
returned by default. This can be changed by setting
integrity_error_status_code. This change is most likely temporary in the
development branch because the uniqueness in the database should probably
be checked some other way(it might also involve locking).